requirements.txt
aiosqlite>=0.17.0
alembic>=1.6.5
asyncpg>=0.23.0
fastapi>=0.65.2
graphene-sqlalchemy>=2.3.0
graphene>=2.1.8
gunicorn>=20.1.0
python-decouple>=3.4
scalecodec>=0.11.13
sqlalchemy>=1.4.19
substrate-interface!=0.13.9
tqdm>=4.61.1
uvicorn>=0.14.0
